Search is a view not an action
